+++ bioptim/__init__.py
@@ -172,7 +172,8 @@ from .dynamics.fatigue.michaud_fatigue import MichaudFatigue, MichaudTauFatigue
 from .dynamics.fatigue.effort_perception import EffortPerception, TauEffortPerception
 from .dynamics.ode_solver import OdeSolver, OdeSolverBase
 from .interfaces.solver_options import Solver
-from .interfaces.biorbd_model import BiorbdModel, MultiBiorbdModel
+from .interfaces.biorbd_model import BiorbdModel
+from .interfaces.multi_biorbd_model import MultiBiorbdModel
 from .interfaces.biomodel import BioModel
 from .interfaces.holonomic_biomodel import HolonomicBioModel
 from .interfaces.variational_biomodel import VariationalBioModel
