+++ bioptim/interfaces/biorbd_model.py
@@ -616,662 +616,3 @@ class BiorbdModel:
             return all_bioviz
 
 
-class MultiBiorbdModel:
-    """
-    This class allows to define multiple biorbd models for the same phase.
-
-
-    Attributes
-    ----------
-    models : list[BiorbdModel]
-        The list of biorbd models to be handled in the optimal control program.
-    extra_models : list[BiorbdModel]
-        A list of extra biorbd models stored in the class for further use.
-
-    Methods
-    -------
-    variable_index()
-        Get the index of the variables in the global vector for a given model index.
-    nb_models()
-        Get the number of models.
-    nb_extra_models()
-        Get the number of extra models.
-
-    """
-
-    def __init__(
-        self,
-        bio_model: tuple[str | biorbd.Model | BiorbdModel, ...],
-        extra_bio_models: tuple[str | biorbd.Model | BiorbdModel, ...] = (),
-    ):
-        self.models = []
-        if not isinstance(bio_model, tuple):
-            raise ValueError("The models must be a 'str', 'biorbd.Model', 'bioptim.BiorbdModel'" " or a tuple of those")
-
-        for model in bio_model:
-            if isinstance(model, str):
-                self.models.append(BiorbdModel(model))
-            elif isinstance(model, biorbd.Model):
-                self.models.append(BiorbdModel(model))
-            elif isinstance(model, BiorbdModel):
-                self.models.append(model)
-            else:
-                raise ValueError("The models should be of type 'str', 'biorbd.Model' or 'bioptim.BiorbdModel'")
-
-        if not isinstance(extra_bio_models, tuple):
-            raise ValueError("The models must be a 'str', 'biorbd.Model', 'bioptim.BiorbdModel'" " or a tuple of those")
-
-        self.extra_models = []
-        for model in extra_bio_models:
-            if isinstance(model, str):
-                self.extra_models.append(BiorbdModel(model))
-            elif isinstance(model, biorbd.Model):
-                self.extra_models.append(BiorbdModel(model))
-            elif isinstance(model, BiorbdModel):
-                self.extra_models.append(model)
-            else:
-                raise ValueError("The models should be of type 'str', 'biorbd.Model' or 'bioptim.BiorbdModel'")
-
-    def __getitem__(self, index):
-        return self.models[index]
-
-    def deep_copy(self, *args):
-        raise NotImplementedError("Deep copy is not implemented yet for MultiBiorbdModel class")
-
-    @property
-    def path(self) -> (list[str], list[str]):
-        return [model.path for model in self.models], [model.path for model in self.extra_models]
-
-    def copy(self):
-        return MultiBiorbdModel(tuple(self.path[0]), tuple(self.path[1]))
-
-    def serialize(self) -> tuple[Callable, dict]:
-        return MultiBiorbdModel, dict(bio_model=tuple(self.path[0]), extra_bio_models=tuple(self.path[1]))
-
-    def variable_index(self, variable: str, model_index: int) -> range:
-        """
-        Get the index of the variables in the global vector for a given model index
-
-        Parameters
-        ----------
-        variable: str
-            The variable to get the index from such as 'q', 'qdot', 'qddot', 'tau', 'contact'
-        model_index: int
-            The index of the model to get the index from
-
-        Returns
-        -------
-        range
-            The index of the variable in the global vector
-        """
-        if variable == "q":
-            current_idx = 0
-            for model in self.models[:model_index]:
-                current_idx += model.nb_q
-            return range(current_idx, current_idx + self.models[model_index].nb_q)
-        elif variable == "qdot":
-            current_idx = 0
-            for model in self.models[:model_index]:
-                current_idx += model.nb_qdot
-            return range(current_idx, current_idx + self.models[model_index].nb_qdot)
-        elif variable == "qddot":
-            current_idx = 0
-            for model in self.models[:model_index]:
-                current_idx += model.nb_qddot
-            return range(current_idx, current_idx + self.models[model_index].nb_qddot)
-        elif variable == "qddot_joints":
-            current_idx = 0
-            for model in self.models[:model_index]:
-                current_idx += model.nb_qddot - model.nb_root
-            return range(
-                current_idx, current_idx + self.models[model_index].nb_qddot - self.models[model_index].nb_root
-            )
-        elif variable == "qddot_root":
-            current_idx = 0
-            for model in self.models[:model_index]:
-                current_idx += model.nb_root
-            return range(current_idx, current_idx + self.models[model_index].nb_root)
-        elif variable == "tau":
-            current_idx = 0
-            for model in self.models[:model_index]:
-                current_idx += model.nb_tau
-            return range(current_idx, current_idx + self.models[model_index].nb_tau)
-        elif variable == "contact":
-            current_idx = 0
-            for model in self.models[:model_index]:
-                current_idx += model.nb_rigid_contacts
-            return range(current_idx, current_idx + self.models[model_index].nb_rigid_contacts)
-
-    @property
-    def nb_models(self) -> int:
-        """
-        Get the number of models
-
-        Returns
-        -------
-        int
-            The number of models
-        """
-        return len(self.models)
-
-    @property
-    def nb_extra_models(self) -> int:
-        """
-        Get the number of extra models
-
-        Returns
-        -------
-        int
-            The number of extra models
-        """
-        return len(self.extra_models)
-
-    @property
-    def gravity(self) -> MX:
-        return vertcat(*(model.gravity for model in self.models))
-
-    def set_gravity(self, new_gravity) -> None:
-        for model in self.models:
-            model.set_gravity(new_gravity)
-        return
-
-    @property
-    def nb_tau(self) -> int:
-        return sum(model.nb_tau for model in self.models)
-
-    @property
-    def nb_segments(self) -> int:
-        return sum(model.nb_segments for model in self.models)
-
-    def segment_index(self, name) -> int:
-        raise NotImplementedError("segment_index is not implemented for MultiBiorbdModel")
-
-    @property
-    def nb_quaternions(self) -> int:
-        return sum(model.nb_quaternions for model in self.models)
-
-    @property
-    def nb_q(self) -> int:
-        return sum(model.nb_q for model in self.models)
-
-    @property
-    def nb_qdot(self) -> int:
-        return sum(model.nb_qdot for model in self.models)
-
-    @property
-    def nb_qddot(self) -> int:
-        return sum(model.nb_qddot for model in self.models)
-
-    @property
-    def nb_root(self) -> int:
-        return sum(model.nb_root for model in self.models)
-
-    @property
-    def segments(self) -> tuple[biorbd.Segment, ...]:
-        out = ()
-        for model in self.models:
-            out += model.segments
-        return out
-
-    def homogeneous_matrices_in_global(self, q, reference_index, inverse=False):
-        raise NotImplementedError("homogeneous_matrices_in_global is not implemented for MultiBiorbdModel")
-
-    def homogeneous_matrices_in_child(self, *args) -> tuple:
-        raise NotImplementedError("homogeneous_matrices_in_child is not implemented for MultiBiorbdModel")
-
-    @property
-    def mass(self) -> MX:
-        return vertcat(*(model.mass for model in self.models))
-
-    def center_of_mass(self, q) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            out = vertcat(out, model.center_of_mass(q_model))
-        return out
-
-    def center_of_mass_velocity(self, q, qdot) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out = vertcat(
-                out,
-                model.center_of_mass_velocity(q_model, qdot_model),
-            )
-        return out
-
-    def center_of_mass_acceleration(self, q, qdot, qddot) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            qddot_model = qddot[self.variable_index("qddot", i)]
-            out = vertcat(
-                out,
-                model.center_of_mass_acceleration(q_model, qdot_model, qddot_model),
-            )
-        return out
-
-    def mass_matrix(self, q) -> list[MX]:
-        out = []
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            out += [model.mass_matrix(q_model)]
-        return out
-
-    def non_linear_effects(self, q, qdot) -> list[MX]:
-        out = []
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out += [model.non_linear_effects(q_model, qdot_model)]
-        return out
-
-    def angular_momentum(self, q, qdot) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out = vertcat(
-                out,
-                model.angular_momentum(q_model, qdot_model),
-            )
-        return out
-
-    def reshape_qdot(self, q, qdot, k_stab=1) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out = vertcat(
-                out,
-                model.reshape_qdot(q_model, qdot_model, k_stab),
-            )
-        return out
-
-    def segment_angular_velocity(self, q, qdot, idx) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out = vertcat(
-                out,
-                model.segment_angular_velocity(q_model, qdot_model, idx),
-            )
-        return out
-
-    @property
-    def name_dof(self) -> tuple[str, ...]:
-        return tuple([dof for model in self.models for dof in model.name_dof])
-
-    @property
-    def contact_names(self) -> tuple[str, ...]:
-        return tuple([contact for model in self.models for contact in model.contact_names])
-
-    @property
-    def nb_soft_contacts(self) -> int:
-        return sum(model.nb_soft_contacts for model in self.models)
-
-    @property
-    def soft_contact_names(self) -> tuple[str, ...]:
-        return tuple([contact for model in self.models for contact in model.soft_contact_names])
-
-    def soft_contact(self, soft_contact_index, *args):
-        current_number_of_soft_contacts = 0
-        out = []
-        for model in self.models:
-            if soft_contact_index < current_number_of_soft_contacts + model.nb_soft_contacts:
-                out = model.soft_contact(soft_contact_index - current_number_of_soft_contacts, *args)
-                break
-            current_number_of_soft_contacts += model.nb_soft_contacts
-        return out
-
-    @property
-    def muscle_names(self) -> tuple[str, ...]:
-        return tuple([muscle for model in self.models for muscle in model.muscle_names])
-
-    @property
-    def nb_muscles(self) -> int:
-        return sum(model.nb_muscles for model in self.models)
-
-    def torque(self, tau_activations, q, qdot) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            tau_activations_model = tau_activations[self.variable_index("tau", i)]
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out = vertcat(
-                out,
-                model.torque(
-                    tau_activations_model,
-                    q_model,
-                    qdot_model,
-                ),
-            )
-        return out
-
-    def forward_dynamics_free_floating_base(self, q, qdot, qddot_joints) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            qddot_joints_model = qddot_joints[self.variable_index("qddot_joints", i)]
-            out = vertcat(
-                out,
-                model.forward_dynamics_free_floating_base(
-                    q_model,
-                    qdot_model,
-                    qddot_joints_model,
-                ),
-            )
-        return out
-
-    def reorder_qddot_root_joints(self, qddot_root, qddot_joints):
-        out = MX()
-        for i, model in enumerate(self.models):
-            qddot_root_model = qddot_root[self.variable_index("qddot_root", i)]
-            qddot_joints_model = qddot_joints[self.variable_index("qddot_joints", i)]
-            out = vertcat(
-                out,
-                model.reorder_qddot_root_joints(qddot_root_model, qddot_joints_model),
-            )
-
-        return out
-
-    def forward_dynamics(self, q, qdot, tau, external_forces=None, f_contacts=None) -> MX:
-        if f_contacts is not None or external_forces is not None:
-            raise NotImplementedError(
-                "External forces and contact forces are not implemented yet for MultiBiorbdModel."
-            )
-
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            tau_model = tau[self.variable_index("tau", i)]
-            out = vertcat(
-                out,
-                model.forward_dynamics(
-                    q_model,
-                    qdot_model,
-                    tau_model,
-                    external_forces,
-                    f_contacts,
-                ),
-            )
-        return out
-
-    def constrained_forward_dynamics(self, q, qdot, tau, external_forces=None) -> MX:
-        if external_forces is not None:
-            raise NotImplementedError("External forces are not implemented yet for MultiBiorbdModel.")
-
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            tau_model = tau[self.variable_index("qddot", i)]  # Due to a bug in biorbd
-            out = vertcat(
-                out,
-                model.constrained_forward_dynamics(
-                    q_model,
-                    qdot_model,
-                    tau_model,
-                    external_forces,
-                ),
-            )
-        return out
-
-    def inverse_dynamics(self, q, qdot, qddot, external_forces=None, f_contacts=None) -> MX:
-        if f_contacts is not None or external_forces is not None:
-            raise NotImplementedError(
-                "External forces and contact forces are not implemented yet for MultiBiorbdModel."
-            )
-
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            qddot_model = qddot[self.variable_index("qddot", i)]
-            out = vertcat(
-                out,
-                model.inverse_dynamics(
-                    q_model,
-                    qdot_model,
-                    qddot_model,
-                    external_forces,
-                    f_contacts,
-                ),
-            )
-        return out
-
-    def contact_forces_from_constrained_forward_dynamics(self, q, qdot, tau, external_forces=None) -> MX:
-        if external_forces is not None:
-            raise NotImplementedError("External forces are not implemented yet for MultiBiorbdModel.")
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            tau_model = tau[self.variable_index("qddot", i)]  # Due to a bug in biorbd
-            out = vertcat(
-                out,
-                model.contact_forces_from_constrained_forward_dynamics(
-                    q_model,
-                    qdot_model,
-                    tau_model,
-                    external_forces,
-                ),
-            )
-        return out
-
-    def qdot_from_impact(self, q, qdot_pre_impact) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_pre_impact_model = qdot_pre_impact[self.variable_index("qdot", i)]
-            out = vertcat(
-                out,
-                model.qdot_from_impact(
-                    q_model,
-                    qdot_pre_impact_model,
-                ),
-            )
-        return out
-
-    def muscle_activation_dot(self, muscle_excitations) -> MX:
-        out = MX()
-        for model in self.models:
-            muscle_states = model.model.stateSet()  # still call from Biorbd
-            for k in range(model.nb_muscles):
-                muscle_states[k].setExcitation(muscle_excitations[k])
-            out = vertcat(out, model.model.activationDot(muscle_states).to_mx())
-        return out
-
-    def muscle_joint_torque(self, activations, q, qdot) -> MX:
-        out = MX()
-        for model in self.models:
-            muscles_states = model.model.stateSet()  # still call from Biorbd
-            for k in range(model.nb_muscles):
-                muscles_states[k].setActivation(activations[k])
-            q_biorbd = GeneralizedCoordinates(q)
-            qdot_biorbd = GeneralizedVelocity(qdot)
-            out = vertcat(out, model.model.muscularJointTorque(muscles_states, q_biorbd, qdot_biorbd).to_mx())
-        return out
-
-    def markers(self, q) -> Any | list[MX]:
-        out = []
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            out.append(model.markers(q_model))
-        return [item for sublist in out for item in sublist]
-
-    @property
-    def nb_markers(self) -> int:
-        return sum(model.nb_markers for model in self.models)
-
-    def marker_index(self, name):
-        raise NotImplementedError("marker_index is not implemented yet for MultiBiorbdModel")
-
-    def marker(self, q, index, reference_segment_index=None) -> MX:
-        raise NotImplementedError("marker is not implemented yet for MultiBiorbdModel")
-
-    @property
-    def nb_rigid_contacts(self) -> int:
-        """
-        Returns the number of rigid contacts.
-        Example:
-            First contact with axis YZ
-            Second contact with axis Z
-            nb_rigid_contacts = 2
-        """
-        return sum(model.nb_rigid_contacts for model in self.models)
-
-    @property
-    def nb_contacts(self) -> int:
-        """
-        Returns the number of contact index.
-        Example:
-            First contact with axis YZ
-            Second contact with axis Z
-            nb_contacts = 3
-        """
-        return sum(model.nb_contacts for model in self.models)
-
-    def rigid_contact_index(self, contact_index) -> tuple:
-        """
-        Returns the axis index of this specific rigid contact.
-        Example:
-            First contact with axis YZ
-            Second contact with axis Z
-            rigid_contact_index(0) = (1, 2)
-        """
-
-        model_selected = None
-        for i, model in enumerate(self.models):
-            if contact_index in self.variable_index("contact", i):
-                model_selected = model
-            # Note: may not work if the contact_index is not in the first model
-        return model_selected.rigid_contact_index(contact_index)
-
-    def marker_velocities(self, q, qdot, reference_index=None) -> list[MX]:
-        if reference_index is not None:
-            raise RuntimeError("marker_velocities is not implemented yet with reference_index for MultiBiorbdModel")
-
-        out = []
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out.extend(
-                model.marker_velocities(q_model, qdot_model, reference_index),
-            )
-        return out
-
-    def tau_max(self, q, qdot) -> tuple[MX, MX]:
-        out_max = MX()
-        out_min = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            torque_max, torque_min = model.tau_max(q_model, qdot_model)
-            out_max = vertcat(out_max, torque_max)
-            out_min = vertcat(out_min, torque_min)
-        return out_max, out_min
-
-    def rigid_contact_acceleration(self, q, qdot, qddot, contact_index, contact_axis) -> MX:
-        model_selected = None
-        model_idx = -1
-        for i, model in enumerate(self.models):
-            if contact_index in self.variable_index("contact", i):
-                model_selected = model
-                model_idx = i
-        q_model = q[self.variable_index("q", model_idx)]
-        qdot_model = qdot[self.variable_index("qdot", model_idx)]
-        qddot_model = qddot[self.variable_index("qddot", model_idx)]
-        return model_selected.rigid_contact_acceleration(q_model, qdot_model, qddot_model, contact_index, contact_axis)
-
-    @property
-    def nb_dof(self) -> int:
-        return sum(model.nb_dof for model in self.models)
-
-    @property
-    def marker_names(self) -> tuple[str, ...]:
-        return tuple([name for model in self.models for name in model.marker_names])
-
-    def soft_contact_forces(self, q, qdot) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            soft_contact_forces = model.soft_contact_forces(q_model, qdot_model)
-            out = vertcat(out, soft_contact_forces)
-        return out
-
-    def reshape_fext_to_fcontact(self, fext: MX) -> biorbd.VecBiorbdVector:
-        raise NotImplementedError("reshape_fext_to_fcontact is not implemented yet for MultiBiorbdModel")
-
-    def normalize_state_quaternions(self, x: MX | SX) -> MX | SX:
-        all_q_normalized = MX()
-        for i, model in enumerate(self.models):
-            q_model = x[self.variable_index("q", i)]  # quaternions are only in q
-            q_normalized = model.normalize_state_quaternions(q_model)
-            all_q_normalized = vertcat(all_q_normalized, q_normalized)
-        idx_first_qdot = self.nb_q  # assuming x = [q, qdot]
-        x_normalized = vertcat(all_q_normalized, x[idx_first_qdot:])
-
-        return x_normalized
-
-    def contact_forces(self, q, qdot, tau, external_forces: list = None) -> MX:
-        if external_forces is not None:
-            raise NotImplementedError("contact_forces is not implemented yet with external_forces for MultiBiorbdModel")
-
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            tau_model = tau[self.variable_index("tau", i)]
-
-            contact_forces = model.contact_forces(q_model, qdot_model, tau_model, external_forces)
-            out = vertcat(out, contact_forces)
-
-        return out
-
-    def passive_joint_torque(self, q, qdot) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out = vertcat(out, model.passive_joint_torque(q_model, qdot_model))
-        return out
-
-    def ligament_joint_torque(self, q, qdot) -> MX:
-        out = MX()
-        for i, model in enumerate(self.models):
-            q_model = q[self.variable_index("q", i)]
-            qdot_model = qdot[self.variable_index("qdot", i)]
-            out = vertcat(out, model.ligament_joint_torque(q_model, qdot_model))
-        return out
-
-    def ranges_from_model(self, variable: str):
-        return [the_range for model in self.models for the_range in model.ranges_from_model(variable)]
-
-    def bounds_from_ranges(self, variables: str | list[str, ...], mapping: BiMapping | BiMappingList = None) -> Bounds:
-        return bounds_from_ranges(self, variables, mapping)
-
-    def _q_mapping(self, mapping: BiMapping = None) -> dict:
-        return _q_mapping(self, mapping)
-
-    def _qdot_mapping(self, mapping: BiMapping = None) -> dict:
-        return _qdot_mapping(self, mapping)
-
-    def _qddot_mapping(self, mapping: BiMapping = None) -> dict:
-        return _qddot_mapping(self, mapping)
-
-    def lagrangian(self):
-        raise NotImplementedError("lagrangian is not implemented yet for MultiBiorbdModel")
-
-    def partitioned_forward_dynamics(self, q_u, qdot_u, tau, external_forces=None, f_contacts=None, q_v_init=None):
-        raise NotImplementedError("partitioned_forward_dynamics is not implemented yet for MultiBiorbdModel")
-
-    @staticmethod
-    def animate(solution: Any, show_now: bool = True, tracked_markers: list = None, **kwargs: Any) -> None | list:
-        raise NotImplementedError("animate is not implemented yet for MultiBiorbdModel")
